Moved shutdown information and variables from WSession to WBbsApp. Changed skey to take an int vs. unsigned char. More ongoing work on removing unneeded unsigned data type usages. Disabled remote.exe and remotes.dat support by default. Fixed display bug in voting booth.

+++ bbs/prot/zmodem.cpp
@@ -686,7 +686,7 @@ int ZProtocol(  ZModem *info )
 	info->timeoutCount = 0;
 	info->noiseCount = 0;
 
-	table = tables[(int)info->state];
+	table = tables[info->state];
 	while( table->type != 99  &&  table->type != info->hdrData[0] )
 	{
 		++table;
